Setter url for macro preview links ut fra context
