+++ src/main/resources/lib/global-values/macro-preview.ts
@@ -3,9 +3,9 @@ import {
     getGlobalValueSet,
     getGvKeyAndContentIdFromUniqueKey,
 } from './global-value-utils';
-import { CONTENT_STUDIO_PATH_PREFIX } from '../constants';
 import { GlobalValueItem } from './types';
 import { CaseTimeUnit } from '../../types/content-types/global-case-time-set';
+import { buildEditorPathFromContext } from '../paths/editor-path';
 
 const enToNo: Record<CaseTimeUnit, string> = {
     days: 'dager',
@@ -48,7 +48,9 @@ export const createGlobalValueMacroPreview = (key: string) => {
         <div>
             <span style='font-size:20px'>${displayName}</span><br/>
             <span style='color:#888888'>${_path}</span><br/>
-            <a href='${CONTENT_STUDIO_PATH_PREFIX}/${contentId}' target='_blank'>[Åpne i editoren]</a><br/>
+            <a href='${buildEditorPathFromContext(
+                contentId
+            )}' target='_blank'>[Åpne i editoren]</a><br/>
             <br/>
             ${buildGlobalValuePreviewString(valueItem)}
         </div>
