+++ src/main/resources/services/service-utils.ts
@@ -1,9 +1,6 @@
 import { Content } from '/lib/xp/content';
-import * as contextLib from '/lib/xp/context';
 import { customSelectorEditIcon } from './custom-selector-icons';
-import { CONTENT_ROOT_PROJECT_ID, CONTENT_STUDIO_PATH_PREFIX } from '../lib/constants';
-import { logger } from '../lib/utils/logging';
-import { getContentProjectIdFromRepoId } from '../lib/utils/repo-utils';
+import { buildEditorPathFromContext } from '../lib/paths/editor-path';
 
 type CustomSelectorUsageHit = {
     name: string;
@@ -32,31 +29,18 @@ const iconWithLink = (href: string, iconData: string) => {
     return `<a href="${href}" class="custom-selector-link">${iconData}</a>`;
 };
 
-const buildEditorUrlFromContext = (contentId: string) => {
-    const { repository } = contextLib.get();
-    if (!repository) {
-        logger.error('Could not determine current repo from context!');
-    }
-
-    const projectId = repository
-        ? getContentProjectIdFromRepoId(repository)
-        : CONTENT_ROOT_PROJECT_ID;
-
-    return `${CONTENT_STUDIO_PATH_PREFIX}/${projectId}/edit/${contentId}`;
-};
-
 // Injects a link in the optional icon field of a customselector hit object
 // (this is almost certainly not an intended usage, but it works :D)
 export const customSelectorHitWithLink = (
     hit: XP.CustomSelectorServiceResponseHit,
     contentId: string
 ): XP.CustomSelectorServiceResponseHit => {
-    const url = buildEditorUrlFromContext(contentId);
+    const editorPath = buildEditorPathFromContext(contentId);
 
     return {
         ...hit,
         icon: {
-            data: iconWithLink(url, hit.icon?.data || customSelectorEditIcon.data),
+            data: iconWithLink(editorPath, hit.icon?.data || customSelectorEditIcon.data),
             type: hit.icon?.type || customSelectorEditIcon.type,
         },
     };
